+++ tests/conftest.py
@@ -17,18 +17,8 @@ def run_start() -> Generator[CliRunner]:
         yield runner
 
 
-@pytest.fixture(scope='module')
-def run_start_module() -> Generator[CliRunner]:
-    runner = CliRunner()
-    with runner.isolated_filesystem():
-        result = runner.invoke(start)
-        assert result.exit_code == 0
-        # Tests code is run here
-        yield runner
-
-
 @pytest.fixture(scope='function')
-def flask_app(run_start_module: CliRunner) -> Flask:  # noqa: ARG001
+def flask_app(run_start: CliRunner) -> Flask:  # noqa: ARG001
     app = site.create_app()
     app.config.update({
         'FLATPAGES_AUTO_RELOAD': True,